+++ src/test/java/sonia/scm/redmine/RedmineIssueTrackerTest.java
@@ -44,6 +44,7 @@ import sonia.scm.template.TemplateType;
 import java.util.Collections;
 
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -69,6 +70,7 @@ public class RedmineIssueTrackerTest {
     redmineIssueTracker = new RedmineIssueTracker(configStore, dataStoreFactory, templateEngineFactory, linkHandlerProvider);
   }
 
+  @Test
   public void shouldWriteGlobalConfigToRedmineConfigStore() {
     RedmineGlobalConfiguration config = new RedmineGlobalConfiguration();
     redmineIssueTracker.setGlobalConfiguration(config);
@@ -84,6 +86,7 @@ public class RedmineIssueTrackerTest {
   @Test
   public void shouldGetRepoConfigurationFromStore() {
     Repository repository = createRepository();
+    when(configStore.getConfiguration()).thenReturn(createValidGlobalConfiguration());
     when(configStore.getConfiguration(repository)).thenReturn(createValidConfiguration());
     redmineIssueTracker.resolveConfiguration(repository);
     verify(configStore).getConfiguration(repository);
@@ -97,6 +100,15 @@ public class RedmineIssueTrackerTest {
     when(configStore.getConfiguration()).thenReturn(createValidGlobalConfiguration());
     redmineIssueTracker.resolveConfiguration(repository);
     verify(configStore).getConfiguration(repository);
+    verify(configStore).getConfiguration();
+  }
+
+  @Test
+  public void shouldReturnGlobalConfigIfRepoConfigIsDisabled() {
+    Repository repository = createRepository();
+    when(configStore.getConfiguration()).thenReturn(createValidGlobalConfigurationWithDisabledRepoConfiguration());
+    redmineIssueTracker.resolveConfiguration(repository);
+    verify(configStore, never()).getConfiguration(repository);
     verify(configStore).getConfiguration();
   }
 
@@ -121,12 +133,23 @@ public class RedmineIssueTrackerTest {
       "user",
       "password");
   }
+
   private RedmineGlobalConfiguration createValidGlobalConfiguration() {
     return new RedmineGlobalConfiguration("http://h2g2.com",
       TextFormatting.MARKDOWN,
       false,
       false,
       false,
+      "user",
+      "password");
+  }
+
+  private RedmineGlobalConfiguration createValidGlobalConfigurationWithDisabledRepoConfiguration() {
+    return new RedmineGlobalConfiguration("http://h2g2.com",
+      TextFormatting.MARKDOWN,
+      false,
+      false,
+      true,
       "user",
       "password");
   }